@@ -2,6 +2,9 @@ package com.example.yazilimguncelkonular.RestApi;
 
 import com.example.yazilimguncelkonular.Models.ResponseReturn;
 import com.example.yazilimguncelkonular.Models.Login;
+import com.example.yazilimguncelkonular.Models.YoklamaResponse;
+
+import java.util.List;
 
 import retrofit2.Call;
 
@@ -12,9 +15,15 @@ public class ManagerAll extends BaseManager{
     {
         return ourInstance;
     }
-    public Call<ResponseReturn> getirCall(Login login)
+    public Call<ResponseReturn> managerAllRestApiLogin(Login login)
     {
-        Call<ResponseReturn> x=getRestApi().getir(login);
+        Call<ResponseReturn> x=getRestApi().restApiLogin(login);
         return x;
     }
+    public Call<List<YoklamaResponse>> managerAllRestApiDers(String token)
+    {
+        Call<List<YoklamaResponse>> x=getRestApi().restApiDers(token);
+        return x;
+    }
+
 }
